// weeb-dashboard/lib/plugins/github/types.d.ts
export interface GithubConfig {
    enabled: boolean;
    sections: string[];
    username: string;
    profile_hide_title?: boolean;
    profile_title?: string;
    profile_hide_avatar?: boolean;
    favorite_languages_title?: string;
    favorite_languages_hide_title?: boolean;
    favorite_license_title?: string;
    favorite_license_hide_title?: boolean;
    favorite_languages_max_languages?: number;
    favorite_languages_ignore_languages?: string;
    repositories_title?: string;
    repositories_hide_title?: boolean;
    repositories_use_private?: boolean;
    activity_title?: string;
    activity_hide_title?: boolean;
    calendar_title?: string;
    calendar_hide_title?: boolean;
    code_habits_title?: string;
    code_habits_hide_title?: boolean;
    code_habits_hide_languages?: boolean;
    code_habits_hide_stats?: boolean;
    code_habits_hide_weekdays?: boolean;
    code_habits_hide_hours?: boolean;
    code_habits_hide_footer?: boolean;
    starred_repositories_title?: string;
    starred_repositories_hide_title?: boolean;
    starred_repositories_max?: number;
    gists_title?: string;
    gists_hide_title?: boolean;
    gists_max?: number;
    stargazers_title?: string;
    stargazers_hide_title?: boolean;
    top_repositories_title?: string;
    top_repositories_hide_title?: boolean;
    top_repositories_max?: number;
    calendar_years?: string;
    star_lists_hide_title?: boolean;
    star_lists_title?: string;
    star_lists_max?: number;
    notable_contributions_hide_title?: boolean;
    notable_contributions_title?: string;
    notable_contributions_max?: number;
    recent_activity_hide_title?: boolean;
    recent_activity_title?: string;
    recent_activity_max?: number;
    introduction_hide_title?: boolean;
    introduction_title?: string;
    introduction_custom_text?: string;
    featured_repositories_hide_title?: boolean;
    featured_repositories_title?: string;
    featured_repositories_urls?: string;
    featured_repositories_max?: number;
    sponsorships_hide_title?: boolean;
    sponsorships_title?: string;
    sponsorships_max?: number;
    sponsors_hide_title?: boolean;
    sponsors_title?: string;
    sponsors_max?: number;
    people_hide_title?: boolean;
    people_title?: string;
    people_type?: 'profile' | 'repository';
    people_repository?: string;
    people_max?: number;
    repository_contributors_hide_title?: boolean;
    repository_contributors_title?: string;
    repository_contributors_repository?: string;
    repository_contributors_max?: number;
}
export interface GithubData {
// ...
    languages: Array<{
        name: string;
        size: number;
        color: string;
    }>;
// ...
}